+++ lib/src/transformers/ignore_elements.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 class _IgnoreElementsStreamSink<S> implements EventSink<S> {
-  final EventSink<S> _outputSink;
+  final EventSink<Never> _outputSink;
 
   _IgnoreElementsStreamSink(this._outputSink);
 
@@ -18,21 +18,25 @@ class _IgnoreElementsStreamSink<S> implements EventSink<S> {
 /// Creates a [Stream] where all emitted items are ignored, only the
 /// error / completed notifications are passed
 ///
+/// [ReactiveX doc](http://reactivex.io/documentation/operators/ignoreelements.html)
+/// [Interactive marble diagram](https://rxmarbles.com/#ignoreElements)
+///
 /// ### Example
 ///
 ///     MergeStream([
 ///       Stream.fromIterable([1]),
 ///       ErrorStream(Exception())
 ///     ])
+///     .transform(IgnoreElementsStreamTransformer())
 ///     .listen(print, onError: print); // prints Exception
-@Deprecated('Use the drain method from the Stream class instead')
-class IgnoreElementsStreamTransformer<S> extends StreamTransformerBase<S, S> {
+class IgnoreElementsStreamTransformer<S>
+    extends StreamTransformerBase<S, Never> {
   /// Constructs a [StreamTransformer] which simply ignores all events from
   /// the source [Stream], except for error or completed events.
   IgnoreElementsStreamTransformer();
 
   @override
-  Stream<S> bind(Stream<S> stream) => Stream.eventTransformed(
+  Stream<Never> bind(Stream<S> stream) => Stream.eventTransformed(
       stream, (sink) => _IgnoreElementsStreamSink<S>(sink));
 }
 
@@ -41,13 +45,17 @@ extension IgnoreElementsExtension<T> on Stream<T> {
   /// Creates a Stream where all emitted items are ignored, only the error /
   /// completed notifications are passed
   ///
+  /// [ReactiveX doc](http://reactivex.io/documentation/operators/ignoreelements.html)
+  /// [Interactive marble diagram](https://rxmarbles.com/#ignoreElements)
+  ///
   /// ### Example
   ///
-  ///    MergeStream([
-  ///      Stream.fromIterable([1]),
-  ///      Stream.error(Exception())
-  ///    ])
-  ///    .listen(print, onError: print); // prints Exception
-  @Deprecated('Use the drain method from the Stream class instead')
-  Stream<T> ignoreElements() => transform(IgnoreElementsStreamTransformer<T>());
+  ///     MergeStream([
+  ///       Stream.fromIterable([1]),
+  ///       Stream<int>.error(Exception())
+  ///     ])
+  ///     .ignoreElements()
+  ///     .listen(print, onError: print); // prints Exception
+  Stream<Never> ignoreElements() =>
+      transform(IgnoreElementsStreamTransformer<T>());
 }