+++ test/transformers/ignore_elements_test.dart
@@ -21,17 +21,40 @@ void main() {
   test('Rx.ignoreElements', () async {
     var hasReceivedEvent = false;
 
-    // ignore: deprecated_member_use_from_same_package
     _getStream().ignoreElements().listen((_) {
       hasReceivedEvent = true;
     },
         onDone: expectAsync0(() {
           expect(hasReceivedEvent, isFalse);
         }, count: 1));
+
+    expect(
+      _getStream().ignoreElements(),
+      emitsInOrder(<Object>[emitsDone]),
+    );
+  });
+
+  test('Rx.ignoreElements.cast', () {
+    final ignored = _getStream().ignoreElements();
+
+    expect(ignored, isA<Stream<void>>());
+    expect(ignored, isA<Stream<Null>>());
+    expect(ignored, isA<Stream<int>>());
+    expect(ignored, isA<Stream<int?>>());
+    expect(ignored, isA<Stream<Object>>());
+    expect(ignored, isA<Stream<Object?>>());
+
+    ignored as Stream<void>; // ignore: unnecessary_cast
+    ignored as Stream<Null>; // ignore: unnecessary_cast
+    ignored as Stream<int>; // ignore: unnecessary_cast
+    ignored as Stream<int?>; // ignore: unnecessary_cast
+    ignored as Stream<Object>; // ignore: unnecessary_cast
+    ignored as Stream<Object?>; // ignore: unnecessary_cast
+
+    expect(true, true);
   });
 
   test('Rx.ignoreElements.reusable', () async {
-    // ignore: deprecated_member_use_from_same_package
     final transformer = IgnoreElementsStreamTransformer<int>();
     var hasReceivedEvent = false;
 
@@ -51,7 +74,6 @@ void main() {
   });
 
   test('Rx.ignoreElements.asBroadcastStream', () async {
-    // ignore: deprecated_member_use_from_same_package
     final stream = _getStream().asBroadcastStream().ignoreElements();
 
     // listen twice on same stream
@@ -64,7 +86,6 @@ void main() {
   test('Rx.ignoreElements.pause.resume', () async {
     var hasReceivedEvent = false;
 
-    // ignore: deprecated_member_use_from_same_package
     _getStream().ignoreElements().listen((_) {
       hasReceivedEvent = true;
     },
@@ -76,7 +97,6 @@ void main() {
   });
 
   test('Rx.ignoreElements.error.shouldThrow', () async {
-    // ignore: deprecated_member_use_from_same_package
     final streamWithError = Stream<void>.error(Exception()).ignoreElements();
 
     streamWithError.listen(null,
@@ -88,7 +108,6 @@ void main() {
   test('Rx.flatMap accidental broadcast', () async {
     final controller = StreamController<int>();
 
-    // ignore: deprecated_member_use_from_same_package
     final stream = controller.stream.ignoreElements();
 
     stream.listen(null);